requirements.txt
transformers
sentence-transformers
openai-clip
pinecone-client
requests
urllib3
gradio
streamlit